requirement.txt
Mindspore
Pillow
opencv-python
numpy